+++ src-old/types/ButtonDescription.ts
@@ -1,5 +1,0 @@
-export default interface ButtonDescription {
-  type: string;
-  pos: Vector;
-  roomIndex: int;
-}