+++ src-old/types/GlobalsRunRoom.ts
@@ -1,5 +1,0 @@
-/*
-export default class GlobalsRunRoom {
-  numDDItems = 0;
-}
-*/