+++ src-old/types/GlobinDescription.ts
@@ -1,5 +1,0 @@
-export interface GlobinDescription {
-  npc: EntityNPC;
-  lastState: int;
-  numRegenerations: int;
-}