+++ src-old/types/HopperDescription.ts
@@ -1,6 +1,0 @@
-export default interface HopperDescription {
-  npc: EntityNPC;
-  posX: float;
-  posY: float;
-  lastMoveFrame: int;
-}