+++ src-old/types/SpriteDescription.ts
@@ -1,4 +1,0 @@
-export default interface SpriteDescription {
-  sprite: Sprite | null;
-  spriteName: string;
-}